bridge/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
python-socketio